Validate positions of var
